rag-application/requirements.txt
# OCR and PDF Processing
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Text Processing and NLP
langchain>=0.1.0
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.30.0
gensim>=4.3.0

# Web Framework
streamlit>=1.28.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Additional utilities
pathlib2>=2.3.7
uuid>=1.30
typing-extensions>=4.7.0

PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
langchain>=0.1.0
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.30.0
gensim>=4.3.0
streamlit>=1.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
uuid>=1.30
typing-extensions>=4.7.0
openai>=1.0.0
qdrant-client>=1.7.0

# Streamlit Demo Requirements
# Add these to your existing requirements.txt

streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0

# Core RAG Pipeline (if not already installed)
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Optional AI Services
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=1.0.0
